+++ src/main/java/com/telesign/MessagingClient.java
@@ -26,9 +26,9 @@ public class MessagingClient extends RestClient {
     public MessagingClient(String customerId,
                            String apiKey,
                            String restEndpoint,
-                           Long connectTimeout,
-                           Long readTimeout,
-                           Long writeTimeout,
+                           Integer connectTimeout,
+                           Integer readTimeout,
+                           Integer writeTimeout,
                            Proxy proxy,
                            final String proxyUsername,
                            final String proxyPassword) {
