+++ src/main/java/com/telesign/PhoneIdClient.java
@@ -24,9 +24,9 @@ public class PhoneIdClient extends RestClient {
     public PhoneIdClient(String customerId,
                          String apiKey,
                          String restEndpoint,
-                         Long connectTimeout,
-                         Long readTimeout,
-                         Long writeTimeout,
+                         Integer connectTimeout,
+                         Integer readTimeout,
+                         Integer writeTimeout,
                          Proxy proxy,
                          final String proxyUsername,
                          final String proxyPassword) {
