+++ src/main/java/com/telesign/ScoreClient.java
@@ -24,9 +24,9 @@ public class ScoreClient extends RestClient {
     public ScoreClient(String customerId,
                        String apiKey,
                        String restEndpoint,
-                       Long connectTimeout,
-                       Long readTimeout,
-                       Long writeTimeout,
+                       Integer connectTimeout,
+                       Integer readTimeout,
+                       Integer writeTimeout,
                        Proxy proxy,
                        final String proxyUsername,
                        final String proxyPassword) {
