+++ src/main/java/com/telesign/VoiceClient.java
@@ -26,9 +26,9 @@ public class VoiceClient extends RestClient {
     public VoiceClient(String customerId,
                        String apiKey,
                        String restEndpoint,
-                       Long connectTimeout,
-                       Long readTimeout,
-                       Long writeTimeout,
+                       Integer connectTimeout,
+                       Integer readTimeout,
+                       Integer writeTimeout,
                        Proxy proxy,
                        final String proxyUsername,
                        final String proxyPassword) {
